requirements.txt
django==4.1.5
django-filter==22.1
djangorestframework==3.14.0
psycopg2-binary==2.9.5
python-environ==0.4.54
Pillow==9.4.0
